Ditch pdfkit for (mostly) home-grown, web-native PDF code.
pdfkit uses a lot of Node.js APIs that require a rather complicated
webpack setup to work on the web. Additionally, it generates a PDF
structure that is not really amenable to linearization, which we
currently don't do, but which we'll probably support in the future,
at least for Writer implementations that support seeking.
One additionally benefit is that we're able to pretty drastically
shrink the bundle size, since we don't include pdfkit, fontkit or
any of the polyfills anymore.