Remove lossless option.
It was severely buggy (as in, the PNGs rendered badly in all PDF viewers)
and I'm running out of time for the christmas release. In addition, it's
a niche feature (who wants gigantic PDFs anyway, and can users really
tell the difference at 300ppi?) that caused quite a bit of bloat in
the dependencies (due to the inclusion of Pako for deflate support).
All in all, I think we're better of without it for the initial release,
and if there's user demand for it we can always come back to it later.