experiment: Fix timeout problem with streamsaver.js
Currently streamsaver will forcefully close the file if no writes have
happened for a while, which can result in prematurely ended PDFS that
are corrupted, e.g. when the remote server takes very long time to
respond to our requests.
This is a small experiment to work around this issue by writing
syntactically irrelevant whitespace to the PDF if no writes have
happened for 20 seconds, in order to prevent this from happening.
Might be removed at a later time if it turns out it doesn't work after
all.