Fix PDF trailer for polyglot ZIP/PDFs to improve compatibility.
Turns out some readers (Acrobat, Edge) expect the crossreference table
to be followed by the beginning of the trailer. Fortunately, the
`startxref...%%EOF` component that other readers rely on to be at
the end of the file does not need to immediately follow the beginning
of the trailer. So by wrapping the ZIP catalog in a dummy PDF object
(that is referenced nowhere, thanks PoC||GTFO#18 for this trick!) we
achieve (hopefully) maximum compatibility with PDF readers and ZIP
unarchivers.