lib: Only use 'top' ranges as top-level ToCs if present (fixes #21)
When converting from IIIFv2 to v3, non-top-level ranges are also added
as entry nodes into the ToC tree when a top-level range is present.
This resulted in tree with duplicate nodes, since each child-level node
would be present twice.
We fix this by making sure that if there's a 'top' range, it will be
selected as the sole top-level ToC node.