Fix infinite loop with certain manifests when estimating PDF size (fixes #23)
When determining the canvases to use for estimating the PDF size, we
checked for the length of all manifests for sampling and used all of
them if there are fewer than the requested number. However, we dont
actually sample from all canvases, but filter for canvases that do not
deviate more than 25% from the mean pixel area of all canvases in the
manifests, to make the estimate more precise.
In some cases, that filtered list of canvases has fewer canvases than
we request for sampling and so the code would eternally try to sample
more.
Instead, we now check the length of the filtered canvases and use all of
them if they're fewer or equal in number to the requested number.