Add support for optimizing JPEG images.
This uses a WASM port of mozjpeg, parallelized via Web Workers, to
recompress the images to reduce the file size while maintaining image
quality as much as possible. Unscientific testing has shown that this
can reduce the file size by up to 50% without much of a quality loss.
As a more performant alternative, re-encoding can also be done with the
browser's JPEG encoder (via Canvas), the results are less impressive
with it, though.

This also switches the build system for the pdiiif library to vite,
since including the web worker in the build was too much of a pain with
the previous tsup build system.